Samples

A data set containing the metadata associated with each sample. Data is obtained from Ruijter (2013), doi:10.1016/j.ymeth.2012.08.011 but original source is by Vermeulen (2009), doi:10.1016/S1470-2045(09)70154-8 .

Usage

samples()

Source

https://medischebiologie.nl/wp-content/uploads/2019/02/qpcrdatamethods.zip

https://github.com/ramiromagno/vermeulen/blob/main/data-raw/samples.csv.gz

Value

A data frame with 372 samples and 3 variables:

sample

Sample identifier.

sample_type

Sample type.

copies

Standard copy number.

dilution

Dilution factor. Higher number means greater dilution.

Details

The tidy version of the data is kept at the repository of the source of {vermeulen} package. This function fetches such data and thus requires internet connection.

See also

Examples

try(samples())
#>         sample sample_type copies dilution
#> 1          111         unk     NA       NA
#> 2         1136         unk     NA       NA
#> 3          116         unk     NA       NA
#> 4         1178         unk     NA       NA
#> 5         1179         unk     NA       NA
#> 6         1231         unk     NA       NA
#> 7          126         unk     NA       NA
#> 8          127         unk     NA       NA
#> 9          130         unk     NA       NA
#> 10        1495         unk     NA       NA
#> 11        1496         unk     NA       NA
#> 12        1497         unk     NA       NA
#> 13        1498         unk     NA       NA
#> 14        1499         unk     NA       NA
#> 15        1500         unk     NA       NA
#> 16        1501         unk     NA       NA
#> 17        1503         unk     NA       NA
#> 18        1507         unk     NA       NA
#> 19        1508         unk     NA       NA
#> 20        1510         unk     NA       NA
#> 21        1511         unk     NA       NA
#> 22        1512         unk     NA       NA
#> 23        1513         unk     NA       NA
#> 24        1514         unk     NA       NA
#> 25        1515         unk     NA       NA
#> 26        1516         unk     NA       NA
#> 27        1517         unk     NA       NA
#> 28        1535         unk     NA       NA
#> 29        1536         unk     NA       NA
#> 30        1538         unk     NA       NA
#> 31        1539         unk     NA       NA
#> 32        1540         unk     NA       NA
#> 33        1541         unk     NA       NA
#> 34        1542         unk     NA       NA
#> 35        1672         unk     NA       NA
#> 36        1677         unk     NA       NA
#> 37        1686         unk     NA       NA
#> 38        1687         unk     NA       NA
#> 39        1693         unk     NA       NA
#> 40        1694         unk     NA       NA
#> 41        1695         unk     NA       NA
#> 42        1696         unk     NA       NA
#> 43        1697         unk     NA       NA
#> 44        1698         unk     NA       NA
#> 45        1699         unk     NA       NA
#> 46        1700         unk     NA       NA
#> 47        1701         unk     NA       NA
#> 48        1702         unk     NA       NA
#> 49        1716         unk     NA       NA
#> 50        1717         unk     NA       NA
#> 51        1720         unk     NA       NA
#> 52        1722         unk     NA       NA
#> 53        1751         unk     NA       NA
#> 54        1752         unk     NA       NA
#> 55        1753         unk     NA       NA
#> 56        1754         unk     NA       NA
#> 57        1755         unk     NA       NA
#> 58        1756         unk     NA       NA
#> 59        1757         unk     NA       NA
#> 60        1758         unk     NA       NA
#> 61        1759         unk     NA       NA
#> 62        1760         unk     NA       NA
#> 63        1761         unk     NA       NA
#> 64        1762         unk     NA       NA
#> 65        1763         unk     NA       NA
#> 66        1764         unk     NA       NA
#> 67        1765         unk     NA       NA
#> 68        1766         unk     NA       NA
#> 69        1767         unk     NA       NA
#> 70        1769         unk     NA       NA
#> 71        1770         unk     NA       NA
#> 72        1771         unk     NA       NA
#> 73        1772         unk     NA       NA
#> 74        1773         unk     NA       NA
#> 75        1774         unk     NA       NA
#> 76        1775         unk     NA       NA
#> 77        1776         unk     NA       NA
#> 78        1777         unk     NA       NA
#> 79        1779         unk     NA       NA
#> 80        1780         unk     NA       NA
#> 81        1781         unk     NA       NA
#> 82        1784         unk     NA       NA
#> 83        1785         unk     NA       NA
#> 84        1786         unk     NA       NA
#> 85        1787         unk     NA       NA
#> 86        1788         unk     NA       NA
#> 87        1789         unk     NA       NA
#> 88        1790         unk     NA       NA
#> 89        1791         unk     NA       NA
#> 90        1792         unk     NA       NA
#> 91        1793         unk     NA       NA
#> 92        1794         unk     NA       NA
#> 93        1795         unk     NA       NA
#> 94        1796         unk     NA       NA
#> 95        1797         unk     NA       NA
#> 96        1798         unk     NA       NA
#> 97        1799         unk     NA       NA
#> 98        1800         unk     NA       NA
#> 99        1801         unk     NA       NA
#> 100       1802         unk     NA       NA
#> 101       1803         unk     NA       NA
#> 102       1804         unk     NA       NA
#> 103       1805         unk     NA       NA
#> 104       1806         unk     NA       NA
#> 105       1810         unk     NA       NA
#> 106       1811         unk     NA       NA
#> 107       1812         unk     NA       NA
#> 108       1813         unk     NA       NA
#> 109       1815         unk     NA       NA
#> 110       1816         unk     NA       NA
#> 111       1818         unk     NA       NA
#> 112       1819         unk     NA       NA
#> 113       1820         unk     NA       NA
#> 114       1821         unk     NA       NA
#> 115       1822         unk     NA       NA
#> 116       1823         unk     NA       NA
#> 117       1824         unk     NA       NA
#> 118       1825         unk     NA       NA
#> 119       1826         unk     NA       NA
#> 120       1827         unk     NA       NA
#> 121       1828         unk     NA       NA
#> 122       1829         unk     NA       NA
#> 123       1830         unk     NA       NA
#> 124       1831         unk     NA       NA
#> 125       1832         unk     NA       NA
#> 126       1834         unk     NA       NA
#> 127       1835         unk     NA       NA
#> 128       1836         unk     NA       NA
#> 129       1837         unk     NA       NA
#> 130       1839         unk     NA       NA
#> 131       1840         unk     NA       NA
#> 132       1841         unk     NA       NA
#> 133       1842         unk     NA       NA
#> 134       1844         unk     NA       NA
#> 135       1846         unk     NA       NA
#> 136       1848         unk     NA       NA
#> 137       1849         unk     NA       NA
#> 138       1850         unk     NA       NA
#> 139       1852         unk     NA       NA
#> 140       1853         unk     NA       NA
#> 141       1854         unk     NA       NA
#> 142       1855         unk     NA       NA
#> 143       1856         unk     NA       NA
#> 144       1857         unk     NA       NA
#> 145       1858         unk     NA       NA
#> 146       1859         unk     NA       NA
#> 147       1860         unk     NA       NA
#> 148       1861         unk     NA       NA
#> 149       1879         unk     NA       NA
#> 150       1880         unk     NA       NA
#> 151       1881         unk     NA       NA
#> 152       1882         unk     NA       NA
#> 153       1883         unk     NA       NA
#> 154       1884         unk     NA       NA
#> 155       1885         unk     NA       NA
#> 156       1886         unk     NA       NA
#> 157       1887         unk     NA       NA
#> 158       1888         unk     NA       NA
#> 159       1889         unk     NA       NA
#> 160       1890         unk     NA       NA
#> 161       1891         unk     NA       NA
#> 162       1892         unk     NA       NA
#> 163       1893         unk     NA       NA
#> 164       1894         unk     NA       NA
#> 165       1895         unk     NA       NA
#> 166       1896         unk     NA       NA
#> 167       1897         unk     NA       NA
#> 168       1898         unk     NA       NA
#> 169       1899         unk     NA       NA
#> 170       1901         unk     NA       NA
#> 171       1902         unk     NA       NA
#> 172       1905         unk     NA       NA
#> 173       1906         unk     NA       NA
#> 174       1907         unk     NA       NA
#> 175       1908         unk     NA       NA
#> 176       1909         unk     NA       NA
#> 177       1910         unk     NA       NA
#> 178       1912         unk     NA       NA
#> 179       1914         unk     NA       NA
#> 180       1915         unk     NA       NA
#> 181       1916         unk     NA       NA
#> 182       1917         unk     NA       NA
#> 183       1918         unk     NA       NA
#> 184       1919         unk     NA       NA
#> 185        192         unk     NA       NA
#> 186       1920         unk     NA       NA
#> 187       1921         unk     NA       NA
#> 188       1922         unk     NA       NA
#> 189       1924         unk     NA       NA
#> 190       1928         unk     NA       NA
#> 191       1929         unk     NA       NA
#> 192        193         unk     NA       NA
#> 193       1931         unk     NA       NA
#> 194       1932         unk     NA       NA
#> 195       1933         unk     NA       NA
#> 196       1934         unk     NA       NA
#> 197       1935         unk     NA       NA
#> 198       1936         unk     NA       NA
#> 199       1937         unk     NA       NA
#> 200       1938         unk     NA       NA
#> 201       1939         unk     NA       NA
#> 202        194         unk     NA       NA
#> 203       1940         unk     NA       NA
#> 204       1941         unk     NA       NA
#> 205       1942         unk     NA       NA
#> 206       1943         unk     NA       NA
#> 207       1944         unk     NA       NA
#> 208       1945         unk     NA       NA
#> 209       1946         unk     NA       NA
#> 210       1947         unk     NA       NA
#> 211       1948         unk     NA       NA
#> 212       1949         unk     NA       NA
#> 213       1950         unk     NA       NA
#> 214       1951         unk     NA       NA
#> 215       1952         unk     NA       NA
#> 216       1953         unk     NA       NA
#> 217       1954         unk     NA       NA
#> 218       1955         unk     NA       NA
#> 219       1956         unk     NA       NA
#> 220       1957         unk     NA       NA
#> 221       1958         unk     NA       NA
#> 222       1959         unk     NA       NA
#> 223       1960         unk     NA       NA
#> 224       1961         unk     NA       NA
#> 225       1962         unk     NA       NA
#> 226       1963         unk     NA       NA
#> 227       1964         unk     NA       NA
#> 228       1965         unk     NA       NA
#> 229       1966         unk     NA       NA
#> 230       1967         unk     NA       NA
#> 231       1968         unk     NA       NA
#> 232       2015         unk     NA       NA
#> 233       2016         unk     NA       NA
#> 234       2017         unk     NA       NA
#> 235       2021         unk     NA       NA
#> 236       2022         unk     NA       NA
#> 237       2025         unk     NA       NA
#> 238       2067         unk     NA       NA
#> 239       2068         unk     NA       NA
#> 240       2069         unk     NA       NA
#> 241       2070         unk     NA       NA
#> 242       2071         unk     NA       NA
#> 243       2084         unk     NA       NA
#> 244       2085         unk     NA       NA
#> 245       2086         unk     NA       NA
#> 246       2087         unk     NA       NA
#> 247       2088         unk     NA       NA
#> 248       2089         unk     NA       NA
#> 249       2090         unk     NA       NA
#> 250       2091         unk     NA       NA
#> 251       2092         unk     NA       NA
#> 252       2093         unk     NA       NA
#> 253       2098         unk     NA       NA
#> 254       2099         unk     NA       NA
#> 255       2100         unk     NA       NA
#> 256       2101         unk     NA       NA
#> 257       2102         unk     NA       NA
#> 258       2103         unk     NA       NA
#> 259       2104         unk     NA       NA
#> 260       2105         unk     NA       NA
#> 261       2106         unk     NA       NA
#> 262       2107         unk     NA       NA
#> 263       2108         unk     NA       NA
#> 264       2110         unk     NA       NA
#> 265       2112         unk     NA       NA
#> 266       2114         unk     NA       NA
#> 267       2115         unk     NA       NA
#> 268       2116         unk     NA       NA
#> 269       2118         unk     NA       NA
#> 270       2119         unk     NA       NA
#> 271       2121         unk     NA       NA
#> 272       2122         unk     NA       NA
#> 273       2159         unk     NA       NA
#> 274       2160         unk     NA       NA
#> 275       2161         unk     NA       NA
#> 276       2162         unk     NA       NA
#> 277       2163         unk     NA       NA
#> 278       2164         unk     NA       NA
#> 279       2165         unk     NA       NA
#> 280       2166         unk     NA       NA
#> 281       2167         unk     NA       NA
#> 282       2168         unk     NA       NA
#> 283       2169         unk     NA       NA
#> 284       2170         unk     NA       NA
#> 285       2171         unk     NA       NA
#> 286       2172         unk     NA       NA
#> 287       2253         unk     NA       NA
#> 288       2257         unk     NA       NA
#> 289       2530         unk     NA       NA
#> 290       2659         unk     NA       NA
#> 291        266         unk     NA       NA
#> 292       2660         unk     NA       NA
#> 293       2661         unk     NA       NA
#> 294       2666         unk     NA       NA
#> 295        267         unk     NA       NA
#> 296       2704         unk     NA       NA
#> 297       2705         unk     NA       NA
#> 298       2706         unk     NA       NA
#> 299       2709         unk     NA       NA
#> 300       2710         unk     NA       NA
#> 301       2711         unk     NA       NA
#> 302       2712         unk     NA       NA
#> 303       2713         unk     NA       NA
#> 304       2714         unk     NA       NA
#> 305       2715         unk     NA       NA
#> 306       2718         unk     NA       NA
#> 307       2719         unk     NA       NA
#> 308       2738         unk     NA       NA
#> 309       2742         unk     NA       NA
#> 310       2743         unk     NA       NA
#> 311       2745         unk     NA       NA
#> 312       2746         unk     NA       NA
#> 313       2748         unk     NA       NA
#> 314       2752         unk     NA       NA
#> 315       2753         unk     NA       NA
#> 316       2754         unk     NA       NA
#> 317       2755         unk     NA       NA
#> 318       2756         unk     NA       NA
#> 319       2757         unk     NA       NA
#> 320       2758         unk     NA       NA
#> 321       2759         unk     NA       NA
#> 322       2760         unk     NA       NA
#> 323       2761         unk     NA       NA
#> 324       2762         unk     NA       NA
#> 325       2763         unk     NA       NA
#> 326       2764         unk     NA       NA
#> 327       2765         unk     NA       NA
#> 328       2768         unk     NA       NA
#> 329       2769         unk     NA       NA
#> 330       2770         unk     NA       NA
#> 331       2771         unk     NA       NA
#> 332       2772         unk     NA       NA
#> 333       2774         unk     NA       NA
#> 334       2775         unk     NA       NA
#> 335       2776         unk     NA       NA
#> 336       2777         unk     NA       NA
#> 337       2778         unk     NA       NA
#> 338       2780         unk     NA       NA
#> 339       2781         unk     NA       NA
#> 340       2782         unk     NA       NA
#> 341       2783         unk     NA       NA
#> 342       2784         unk     NA       NA
#> 343       2785         unk     NA       NA
#> 344       2786         unk     NA       NA
#> 345       2788         unk     NA       NA
#> 346       2793         unk     NA       NA
#> 347       2794         unk     NA       NA
#> 348       2795         unk     NA       NA
#> 349       2796         unk     NA       NA
#> 350        315         unk     NA       NA
#> 351        318         unk     NA       NA
#> 352        320         unk     NA       NA
#> 353        321         unk     NA       NA
#> 354        323         unk     NA       NA
#> 355        330         unk     NA       NA
#> 356        332         unk     NA       NA
#> 357        334         unk     NA       NA
#> 358        338         unk     NA       NA
#> 359        346         unk     NA       NA
#> 360        520         unk     NA       NA
#> 361        764         unk     NA       NA
#> 362        769         unk     NA       NA
#> 363        772         unk     NA       NA
#> 364        824         unk     NA       NA
#> 365        827         unk     NA       NA
#> 366        939         unk     NA       NA
#> 367     STD_15         std     15    10000
#> 368    STD_150         std    150     1000
#> 369   STD_1500         std   1500      100
#> 370  STD_15000         std  15000       10
#> 371 STD_150000         std 150000        1
#> 372      water         ntc      0      Inf