Targets

A data set containing the metadata associated with each target. Data is obtained from Ruijter (2013), doi:10.1016/j.ymeth.2012.08.011 but original source is by Vermeulen (2009), doi:10.1016/S1470-2045(09)70154-8 .

Usage

targets()

Source

https://medischebiologie.nl/wp-content/uploads/2019/02/qpcrdatamethods.zip

https://github.com/ramiromagno/vermeulen/blob/main/data-raw/targets.csv.gz

Value

A data frame with 64 targets and two variables:

target

Target identifier, in almost all cases the name of a gene.

target_type

Target type: either target of interest ("toi") or reference target ("ref").

Details

The tidy version of the data is kept at the repository of the source of {vermeulen} package. This function fetches such data and thus requires internet connection.

See also

Examples

try(targets())
#>     target target_type
#> 1   CAMTA2         toi
#> 2   CAMTA1         toi
#> 3    BIRC5         toi
#> 4  ARHGEF7         toi
#> 5    ALUsq         ref
#> 6   AKR1C1         toi
#> 7     AHCY         toi
#> 8    ECEL1         toi
#> 9   DPYSL3         toi
#> 10     DDC         toi
#> 11   CPSG3         toi
#> 12  CLSTN1         toi
#> 13   CDKN3         toi
#> 14    CDH5         toi
#> 15   CDCA5         toi
#> 16    CD44         toi
#> 17   MRPL3         toi
#> 18    MCM2         toi
#> 19    MAPT         toi
#> 20    MAP7         toi
#> 21  MAP2K4         toi
#> 22   INPP1         toi
#> 23   IGSF4         toi
#> 24   HPRT1         ref
#> 25    HMBS         ref
#> 26  HIVEP2         toi
#> 27    GNB1         toi
#> 28     FYN         toi
#> 29    EPN2         toi
#> 30   EPHA5         toi
#> 31 EPB41L3         toi
#> 32  ELAVL4         toi
#> 33  PRKACB         toi
#> 34   PRDM2         toi
#> 35   PRAME         toi
#> 36   PMP22         toi
#> 37    PLAT         toi
#> 38  PLAGL1         toi
#> 39  PIK3R1         toi
#> 40 PDE4DIP         toi
#> 41   PAICS         toi
#> 42    ODC1         toi
#> 43   NTRK1         toi
#> 44   NRCAM         toi
#> 45   NM23A         toi
#> 46   NHLH2         toi
#> 47    MYCN         toi
#> 48   MTSS1         toi
#> 49    WSB1         toi
#> 50    ULK2         toi
#> 51     UBC         ref
#> 52    TYMS         toi
#> 53  TNFRSF         toi
#> 54  SNAPC1         toi
#> 55 SLC25A5         toi
#> 56  SLC6A8         toi
#> 57    SDHA         ref
#> 58    SCG2         toi
#> 59    QPCT         toi
#> 60  PTPRN2         toi
#> 61   PTPRH         toi
#> 62   PTPRF         toi
#> 63     PTN         toi
#> 64   PRKCZ         toi